refactor file locations in preparation for merge
