+++ src/GameCards/GameCard.css
@@ -59,6 +59,8 @@
   display: flex;
   align-items: flex-start;
   background-color: rgb(24, 23, 23);
+  transition: 2s;
+
 }
 
 @font-face {
